experiment/pyproject.toml
[tool.poetry]
name = "lie-detector"
version = "0.1.0"
description = ""
authors = ["KN Neuron"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pygame = "^2.5.2"
pyyaml = "^6.0.1"
brainaccess = {git = "https://gitlneural_networks.com/kn-neuron/brainaccess-python-api.git"}
jupyter = "^1.0.0"
prettytable = "^3.10.0"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
